services/qr-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
qrcode[pil]==7.4.2
Pillow==10.0.1